debugging why environment variable lookup fails in some cases.

+++ c/f2_ptypes.c
@@ -884,7 +884,7 @@ f2ptr f2string__primobject_type__new(f2ptr cause) {
 
 // symbol
 
-f2ptr ptype_symbol__new(int pool_index, f2ptr cause, uint length, u8* str) {
+f2ptr ptype_symbol__new(int pool_index, f2ptr cause, u64 length, u8* str) {
   if (length == 0) {
     return nil;
   }
@@ -2008,10 +2008,10 @@ void funk2_symbol_hash__add_symbol(funk2_symbol_hash_t* this, f2ptr symbol_f2ptr
   this->array[bin_index] = new_node;
 }
 
-f2ptr funk2_symbol_hash__lookup_symbol__thread_unsafe(funk2_symbol_hash_t* this, uint length, u8* str) {
+f2ptr funk2_symbol_hash__lookup_symbol__thread_unsafe(funk2_symbol_hash_t* this, u64 length, u8* str) {
   ptype_symbol_block_t* symbol_block = NULL;
   // search for chararray in hashed symbols
-  uint bin_index = (uint)((uint)chararray__hash_value(length, str) & (uint)(this->hash_value_bit_mask));
+  u64 bin_index = (u64)((u64)chararray__hash_value(length, str) & (u64)(this->hash_value_bit_mask));
   funk2_symbol_hash_node_t* node = (this->array[bin_index]);
   while (node) {
     symbol_block = (ptype_symbol_block_t*)from_ptr(f2ptr_to_ptr(node->symbol));
@@ -2026,14 +2026,14 @@ f2ptr funk2_symbol_hash__lookup_symbol__thread_unsafe(funk2_symbol_hash_t* this,
   return nil;
 }
 
-f2ptr funk2_symbol_hash__lookup_symbol(funk2_symbol_hash_t* this, uint length, u8* str) {
+f2ptr funk2_symbol_hash__lookup_symbol(funk2_symbol_hash_t* this, u64 length, u8* str) {
   funk2_processor_mutex__user_lock(&this->mutex);
   f2ptr result = funk2_symbol_hash__lookup_symbol__thread_unsafe(this, length, str);
   funk2_processor_mutex__unlock(&(this->mutex));
   return result;
 }
 
-f2ptr funk2_symbol_hash__lookup_or_create_symbol__thread_unsafe(funk2_symbol_hash_t* this, int pool_index, f2ptr cause, uint length, u8* str) {
+f2ptr funk2_symbol_hash__lookup_or_create_symbol__thread_unsafe(funk2_symbol_hash_t* this, int pool_index, f2ptr cause, u64 length, u8* str) {
   f2ptr symbol_f2ptr = funk2_symbol_hash__lookup_symbol__thread_unsafe(this, length, str);
   if (symbol_f2ptr) {
     return symbol_f2ptr;
@@ -2049,14 +2049,14 @@ f2ptr funk2_symbol_hash__lookup_or_create_symbol__thread_unsafe(funk2_symbol_has
   if (str) {memcpy(symbol_block->str, str, length);}
   else     {bzero(symbol_block->str, length);}
   symbol_block->str[length] = 0x00;
-  symbol_block->hash_value  = (u64)((uint)chararray__hash_value(length, str));
+  symbol_block->hash_value  = (u64)((u64)chararray__hash_value(length, str));
   
   // and add new symbol to hash table
   funk2_symbol_hash__add_symbol(this, symbol_f2ptr);
   return symbol_f2ptr;
 }
 
-f2ptr funk2_symbol_hash__lookup_or_create_symbol(funk2_symbol_hash_t* this, int pool_index, f2ptr cause, uint length, u8* str) {
+f2ptr funk2_symbol_hash__lookup_or_create_symbol(funk2_symbol_hash_t* this, int pool_index, f2ptr cause, u64 length, u8* str) {
   funk2_processor_mutex__user_lock(&this->mutex);
   f2ptr result = funk2_symbol_hash__lookup_or_create_symbol__thread_unsafe(this, pool_index, cause, length, str);
   funk2_processor_mutex__unlock(&(this->mutex));
